+++ 2024 Code/src/main/java/frc/robot/util/SubsystemContainer.java
@@ -11,11 +11,11 @@ import frc.robot.subsystems.simulation.Climb;
 import frc.robot.subsystems.simulation.IntakeSimulation;
 import frc.robot.subsystems.simulation.ArmSimulation;
 import frc.robot.subsystems.simulation.Shooter;
-import frc.robot.subsystems.simulation.Swerve;
+import frc.robot.subsystems.simulation.SwerveSim;
 
 public class SubsystemContainer {
     public static final LimelightSubsystem limelightSubsystem = new LimelightSubsystem();
-    public static final Swerve swerveSubsystem = new Swerve();
+    public static final SwerveSim swerveSubsystem = new SwerveSim();
     public static final IntakeSimulation intakeSubsystem = new IntakeSimulation();
     public static final Shooter ShooterSubsystem = new Shooter();
     public static final Climb climbSubsystem = new Climb();
